モデルの指定方法の変更に対応
